+++ app/sprinkles/account/src/Authenticate/Authenticator.php
@@ -20,7 +20,7 @@ use UserFrosting\Sprinkle\Account\Authenticate\Exception\AuthCompromisedExceptio
 use UserFrosting\Sprinkle\Account\Authenticate\Exception\AuthExpiredException;
 use UserFrosting\Sprinkle\Account\Authenticate\Exception\InvalidCredentialsException;
 use UserFrosting\Sprinkle\Account\Database\Models\User;
-use UserFrosting\Sprinkle\Account\Util\Password;
+use UserFrosting\Sprinkle\Account\Facades\Password;
 use UserFrosting\Sprinkle\Core\Util\ClassMapper;
 
 /**