+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## v4.1.15-alpha
+- Refactor `Password` into a instantiable `Hasher` class, service, and `Password` facade (#827)
 - Change default hash cost back to 10 and fix legacy hash detection issue
 
 ## v4.1.14-alpha